+++ nts.uk/uk.at/at.ctx/record/nts.uk.ctx.at.record.infra/src/main/java/nts/uk/ctx/at/record/infra/entity/daily/anyitem/KrcdtDayAnyItemValue.java
@@ -1,6 +1,7 @@
 package nts.uk.ctx.at.record.infra.entity.daily.anyitem;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -71,13 +72,13 @@ public class KrcdtDayAnyItemValue extends UkJpaEntity implements Serializable{
 			this.timeValue = v.valueAsMinutes();
 		});
 		anyItemValue.getTimes().ifPresent(v -> {
-			this.countValue = v.v();
+			this.countValue = v.v().intValue();
 		});
 	}
 	
 	public AnyItemValue toDomain() {
 		return new AnyItemValue(new AnyItemNo(krcdtDayAnyItemValuePK.itemNo),
-								Optional.of(new AnyItemTimes(countValue)), 
+								Optional.of(new AnyItemTimes(BigDecimal.valueOf(countValue))), 
 								Optional.of(new AnyItemAmount(moneyValue)),
 								Optional.of(new AnyItemTime(timeValue)));
 	}
